refactor(shared-types): split index into topical modules with a thin barrel (F-754)
