+++ packages/shared-types/test/export-surface.test.ts
@@ -1,0 +1,159 @@
+// SPDX-License-Identifier: Apache-2.0
+//
+// F-754 export-surface guard. The `@pome-sh/shared-types` barrel is the public
+// V1 contract; consumers (CLI, twins, pome-cloud) import runtime values by name
+// from it. The topical-module split (F-754) must be provably zero-surface-change:
+// if a re-export is dropped, renamed, or shadowed by an `export *` collision,
+// the sorted key list below drifts and this test fails loud.
+//
+// The expected list is an explicit inline array (not a `.snap` file) so the
+// public surface reads plainly in review. It was generated from the pre-refactor
+// `main` build. When you intentionally add/remove an export, update this list in
+// the same PR.
+
+import { describe, expect, it } from "vitest";
+import * as api from "../src/index.js";
+
+// Runtime value exports (types are erased and cannot appear on `Object.keys`).
+const EXPECTED_EXPORTS = [
+  "CRITERION_KINDS",
+  "ERROR_TYPE",
+  "GEN_AI_AGENT_ID",
+  "GEN_AI_AGENT_NAME",
+  "GEN_AI_OPERATION_NAME",
+  "GEN_AI_PROVIDER_NAME",
+  "GEN_AI_REQUEST_MODEL",
+  "GEN_AI_SYSTEM_DEPRECATED",
+  "GEN_AI_TOOL_NAME",
+  "GEN_AI_USAGE_COMPLETION_TOKENS_LEGACY",
+  "GEN_AI_USAGE_INPUT_TOKENS",
+  "GEN_AI_USAGE_OUTPUT_TOKENS",
+  "GEN_AI_USAGE_PROMPT_TOKENS_LEGACY",
+  "GITHUB_ACCESS_CONTROL_CATALOG",
+  "GITHUB_ACCESS_CONTROL_CATEGORY_LABELS",
+  "GITHUB_ACCESS_CONTROL_CATEGORY_ORDER",
+  "HTTP_REQUEST_METHOD",
+  "HTTP_RESPONSE_STATUS_CODE",
+  "KNOWN_TWIN_IDS",
+  "LEGACY_ATTR_NAMESPACE",
+  "LEGACY_CRITERION_KIND_MAP",
+  "LEGACY_ID_PREFIX",
+  "LEGACY_SHIM_SEMCONV_VERSION",
+  "LEGACY_TASK_VOCAB_KEY_MAP",
+  "MOUNTED_TWINS",
+  "OTEL_CORE_SEMCONV_VERSION",
+  "OTEL_GENAI_SCHEMA_URL",
+  "OTEL_GENAI_SCHEMA_VERSION",
+  "OTEL_PROJECTION_KEYS",
+  "OTEL_SPAN_KINDS",
+  "OTEL_STATUS_CODES",
+  "SERVER_ADDRESS",
+  "SERVER_PORT",
+  "UINT64_MAX",
+  "URL_FULL",
+  "URL_PATH",
+  "acceptInviteRequestSchema",
+  "acceptInviteResponseSchema",
+  "apiErrorSchema",
+  "apiErrorTypeSchema",
+  "apiKeyCreatedSchema",
+  "apiKeySchema",
+  "canonicalSpanIdSchema",
+  "canonicalTraceIdSchema",
+  "compareUint64",
+  "correlatorKindSchema",
+  "createEvalSessionResponseSchema",
+  "createInviteRequestSchema",
+  "createInviteResponseSchema",
+  "createSessionRequestSchema",
+  "createSessionResponseSchema",
+  "criterionDefSchema",
+  "criterionKindSchema",
+  "criterionResultSchema",
+  "criterionSchema",
+  "deterministicCriterionResultSchema",
+  "eventSchema",
+  "finalizeAcceptedResponseSchema",
+  "finalizeCompletedStatusResponseSchema",
+  "finalizeFailedStatusResponseSchema",
+  "finalizeFailureErrorSchema",
+  "finalizeInitialResponseSchema",
+  "finalizeQueuedStatusResponseSchema",
+  "finalizeResponseSchema",
+  "finalizeRunningStatusResponseSchema",
+  "finalizeStatusResponseSchema",
+  "finalizeStatusUrlSchema",
+  "formatGitHubAccessControlLabel",
+  "githubAccessControlCatalogSchema",
+  "githubAccessControlCategoryGroupSchema",
+  "githubAccessControlCategorySchema",
+  "githubAccessControlEndpointSchema",
+  "githubAccessControlToolNames",
+  "githubSeedStateSchema",
+  "groupGitHubAccessControlByCategory",
+  "hookEventSchema",
+  "isLegacyEventRow",
+  "isUint64",
+  "judgeModelSchema",
+  "laneSchema",
+  "llmCallEventSchema",
+  "mapOtelSpanToEvent",
+  "meResponseSchema",
+  "msToNanos",
+  "nanosToIso",
+  "normalizeTaskVocabKeys",
+  "otelAttributeValueSchema",
+  "otelEventSchema",
+  "otelSpanEventSchema",
+  "otelSpanInputSchema",
+  "otelSpanKindSchema",
+  "otelStatusCodeSchema",
+  "persistedScenarioSchema",
+  "persistedTaskSchema",
+  "planTierSchema",
+  "probabilisticCriterionResultSchema",
+  "projectAttributes",
+  "providerScopedSeedStateSchema",
+  "recorderEventSchema",
+  "recorderFidelitySchema",
+  "redactEvent",
+  "redactSecrets",
+  "runSchema",
+  "scenarioConfigSchema",
+  "scenarioSchema",
+  "seedStateSchema",
+  "sessionPublicSchema",
+  "sessionSchema",
+  "sessionStateSchema",
+  "shimLegacyEventToSpan",
+  "shimmableLegacyEventSchema",
+  "slackSeedStateSchema",
+  "stateDeltaSchema",
+  "stepSchema",
+  "stripeSeedStateSchema",
+  "subagentSpawnEventSchema",
+  "submitResultRequestSchema",
+  "submitResultResponseSchema",
+  "summarizeGitHubAccessControlCatalog",
+  "taskConfigSchema",
+  "taskSchema",
+  "teamInviteSchema",
+  "teamMemberSchema",
+  "teamRoleSchema",
+  "teamSchema",
+  "toolResultEventSchema",
+  "toolUseEventSchema",
+  "twinHttpEventSchema",
+  "twinIdSchema",
+  "unixNanoSchema",
+  "usageResponseSchema",
+  "userSchema",
+  "w3cSpanIdSchema",
+  "w3cTraceIdSchema",
+] as const;
+
+describe("@pome-sh/shared-types barrel export surface (F-754)", () => {
+  it("re-exports exactly the pre-refactor runtime value surface", () => {
+    expect(Object.keys(api).sort()).toEqual([...EXPECTED_EXPORTS]);
+  });
+});
